Allow deployment to envs without migration queue
Deploying the service to environments that don't currently have a
migration queue is failing because of a failure to parse missing
environment variables.

This change adds an enabled flag for the migration queue for bean
generation, which defaults to enabled in an local environment. Default
values of empty and false (for enabled) are passed when deploying to
environments without a queue.

An additional flag is specified to `kd` to not fail on missing
variables, resulting in no values being passed.